Made Required changes in Player and Scoreboard Table Made Player Endpoints more refined, now it can take list of players

+++ src/main/java/com/tekion/GameOfCricket/Entity/MatchEntity.java
@@ -10,8 +10,8 @@ public class MatchEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private int firstTeamID;
-    private int secondTeamID;
+    private Long firstTeamID;
+    private Long secondTeamID;
     private String winner;
     private int numberOfOvers;
 }
